Allow cons cell style for mode line padding
The 'modus-themes-mode-line' can accept a padding value property in the
form of a natural number.  For example:

    (setq modus-themes-mode-line '(accented borderless 4))

With this addition, the padding can be expressed as a cons cell which
may help with its clarity (otherwise you might forget what the number is
supposed to reference).

These are all the same:

    (setq modus-themes-mode-line '(accented borderless (padding . 4)))
    (setq modus-themes-mode-line '(accented borderless (padding 4)))
    (setq modus-themes-mode-line '(accented borderless 4))

Thanks to Daniel Mendler (else @minad) for proposing this idea in issue
282: <https://gitlab.com/protesilaos/modus-themes/-/issues/282#note_842257619>.

It will be implemented for all relevant user options.