Add optional height to modus-themes-mode-line
The 'modus-themes-mode-line' can now accept a height value in the form
of a floating point.  This represents a rate relative to the main font
size.  For example:

    (setq modus-themes-mode-line '(accented borderless 0.9))

Apart from a floating point, the height can be expressed as a cons cell
in the form of '(height . FLOAT)' or '(height 0.9)'.

These are all the same:

    (setq modus-themes-mode-line '(accented borderless (height . 0.9)))
    (setq modus-themes-mode-line '(accented borderless (height 0.9)))
    (setq modus-themes-mode-line '(accented borderless 0.9))

Also see commit 1706af0 which adjusts the padding property accordingly.

Putting it all together (these are all the same):

    (setq modus-themes-mode-line '(accented borderless (padding . 4) (height . 0.9)))
    (setq modus-themes-mode-line '(accented borderless (padding 4) (height 0.9)))
    (setq modus-themes-mode-line '(accented borderless 4 0.9))

Thanks to Daniel Mendler (else @minad) for proposing this idea in issue
282: <https://gitlab.com/protesilaos/modus-themes/-/issues/282#note_842257619>.