Allow cons cell value for headings' height
This is the same principle as with commits ecc71f6, 1706af0.

The 'modus-themes-headings' accepts a height property either as a
floating point or a cons cell in the form of '(height . FLOAT)' or
'(height FLOAT)'.

Sample with various ways of expressing the height value alongside other
properties (read the doc string or the manual):

    (setq modus-themes-headings
          '((1 . (variable-pitch light (height . 1.6)))
            (2 . (rainbow regular (height 1.4)))
            (3 . (overline background regular 1.3))
            (4 . (1.2))
            (5 . (1.1))
            (t . (monochrome 1.05))))

Same but with a simpler alist notation (this is an Emacs Lisp design,
not specific to the themes):

    (setq modus-themes-headings ; this is an alist: read the manual or its doc string
          '((1 variable-pitch light (height . 1.6))
            (2 rainbow regular (height 1.4))
            (3 overline background regular 1.3)
            (4 1.2)
            (5 1.1)
            (t monochrome 1.05)))