Allow cons cell value for box button height
This is the same principle as with commits ecc71f6, 1706af0, 33df824.

The 'modus-themes-box-buttons' accepts a height property either as a
floating point or a cons cell in the form of '(height . FLOAT)' or
'(height FLOAT)'.

Example with the various ways to specify the height:

    (setq modus-themes-box-buttons '(flat semibold 0.9))
    (setq modus-themes-box-buttons '(flat semibold (height 0.9)))
    (setq modus-themes-box-buttons '(flat semibold (height . 0.9)))

These all give the same height.