Make the Avy "grey" face work better
This face appears when you are asked to type a sequence of characters.
Each character you type is rendered in this face.

The face used to be gray and bold before.  The bold weight made it look
like a valid candidate, while the gray did not contrast well with the
intense magenta.

By removing the bold weight, we make it clear that the character is no
longer a valid target (it was already inserted), while the "cold"
colours contrast better with intense magenta and the other backgrounds.