Review Avy colours for better results
I tried various styles, such as:

    (setq avy-style 'pre)
    (setq avy-style 'at-full)

Also modified the 'avy-lead-faces' (which is a 'defconst', not a
'defcustom'):

    (setq avy-lead-faces
          '(avy-lead-face
            avy-lead-face-0
            avy-lead-face-2
            avy-lead-face
            avy-lead-face-0
            avy-lead-face-2))

    (setq avy-lead-faces
          '(avy-lead-face
            avy-lead-face-1
            avy-lead-face-1
            avy-lead-face-1
            avy-lead-face-1))

    (setq avy-lead-faces
          '(avy-lead-face
            avy-lead-face-2
            avy-lead-face-2
            avy-lead-face-2
            avy-lead-face-2))

I feel these new colour combinations work better than they did before.
Please let me know if this is not the case in your particular workflow
and I am willing to review things.