Reify themes with eval-and-compile
This is an attempt to fix a bug that has existed since version 1.2.0 of
the themes or even earlier.

The bug is about a mismatch between compiled code and runtime
dependencies.  The runtime expects the current version while the
compiled code only furnishes an outdated one, thus resulting in an
error.  This only happens when:

1. Private functions change to accept more/fewer arguments.
2. Variables change their acceptable value (e.g. from symbol to list).
3. The user is installing the package via the package.el mechanism which
   takes care of byte compilation (though anything that mimics
   package.el should exhibit the same behaviour).

My understanding is that the cause was the limited scope of the
'eval-and-compile' we had before: it would run the 'require' also at
compile time, whereas the 'modus-themes-theme' macro, which reifies the
actual theme, would only be evaluated at runtime.  Hence the mismatch as
'require' would read the already installed byte code while the macro
would expect newer forms.

Wrapping everything in the 'eval-and-compile' should address this
problem.  Hopefully it will not engender new ones...

* * *

The latest reports about this bug:

* GitLab issue 287 with Mark Bestley and Daniel Mendler:
  <https://gitlab.com/protesilaos/modus-themes/-/issues/287>.

* GitHub issue 22 with Rytis Paškauskas:
  <https://github.com/protesilaos/modus-themes/issues/22>.