Make generic widget buttons look like box buttons
This means that their looks are controlled by the user option
'modus-themes-box-buttons'.

Note that by default these buttons are wrapped in square brackets,
subject to user options:

* widget-link-prefix
* widget-link-suffix
* widget-push-button-prefix
* widget-push-button-suffix

The brackets introduce a faux box effect on either side which makes it
awkward to apply a genuine box attribute to the applicable faces.  We
normally do not change user options at the theme level, though this is
one of the cases where it is justified to do so.

Thanks to Daniel Mendler for the feedback in issue 296:
<https://gitlab.com/protesilaos/modus-themes/-/issues/296>.