Optionally make generic widget buttons boxy
This concerns the 'modus-themes-box-buttons' user option.  By default,
it will not affect the faces of the widget library.  Those are used, for
example, in the Notmuch "hello" buffer.

Adding the 'all-buttons' property to the aforementioned user option will
make those generic faces use the same style as the Custom UI and
related.

Thanks to Daniel Mendler, Rudolf Adamkovič, and Tony Zorman for their
feedback in issue 296:
<https://gitlab.com/protesilaos/modus-themes/-/issues/296>.