Revert commits about compilation/grep numbers
This reverts commits aafd02640a6ea194badf44fa54f8a55880e80089 and
98faf9b50a8216f3b911aa7f29cc16b41c44a279.

After receiving feedback from Daniel Mendler I understand that the
problems reported by Rudolf Adamkovič in issue 277 are situational and
not the norm with compilation buffers:
<https://gitlab.com/protesilaos/modus-themes/-/issues/277>.

My hesitation and thus my inability to make the correct judgement call
was due to the fact that I do not have experience with compilation
buffers: I only use them in the context of 'M-x grep' and related, where
the old design looked perfectly fine.

Issue 277 has been ongoing since early February 2022, so I thought the
only way to elicit feedback was to make the now-reverted changes and see
what happens.  I did so begrudgingly.  My apologies for this
back-and-forth!