Make 'line-number' follow modus-themes-mixed-fonts
This means that if the user option 'modus-themes-mixed-fonts' is
non-nil, line numbers of 'display-line-numbers-mode' will use a
monospaced typeface at all times (the 'fixed-pitch' face, as explained
in the Modus themes' manual).  Otherwise they use whatever font the
'default' face has.  This makes it better when the user enables
'variable-pitch-mode' but still wants spacing-sensitive constructs to
remain monospaced.

Thanks to Christopher League for the feedback in issue 302:
<https://gitlab.com/protesilaos/modus-themes/-/issues/302>.