Move official Git repository to SourceHut
The new source is <https://git.sr.ht/~protesilaos/modus-themes>.

The GitLab URL which once was the main source of this project will
function as a mirror from now on.  This is in addition to the GitHub
equivalent:

* <https://gitlab.com/protesilaos/modus-themes/>
* <https://github.com/protesilaos/modus-themes/>

The canonical mailing list for user feedback and/or patches is
<https://lists.sr.ht/~protesilaos/modus-themes>.  Everyone is encouraged
to use it: it is ordinary email and one can attach files (e.g. pictures)
if needed.