Add support for calibredb
I have tried to limit the wanton use of colour in the relevant buffers
and also align the package with the overall style of the themes.

The currently selected line is affected by the user option
'modus-themes-hl-line'.

Thanks to Ivan Popovych for the feedback over at the official mailing
list:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C87zgkgroi7.fsf%40gmail.com%3E>.