Stop using 'inverse-video' in powerline
We now apply the colours directly.  The reason is that 'inverse-video'
makes it tricky to override the face as it swaps the foreground with the
background.  That behaviour is only needed in special cases: powerline
is not one of them.

Thanks to Thibaut Verron for the feedback in issue 305 over at the
GitLab mirror: <https://gitlab.com/protesilaos/modus-themes/-/issues/305>.