Make git commit/rebase comments inherit font-lock
This makes it possible to customise 'font-lock-comment-face' and have
the changes apply to those faces as well.  Such a customisation can, for
example, involve the change of the font family or the addition of a
background colour.  We want the whole comment block, including those
special keywords from Git, to look consistent.

This change also makes git-{commit,rebase}-comment-heading attain the
foreground colour of comments, instead of the default one (black or
white).  It makes it look part of the comment block.