Disable byte compilation
Commit f067d2e explains at length the rationale for wrapping the
reification of the themes in 'eval-and-compile'.  While that appears to
solve the problem documented therein, it creates a new one: when we try
to byte compile a theme, we get:

    Error: Lisp nesting exceeds ‘max-lisp-eval-depth’

I think byte compilation is nice to have, but it keeps giving us
problems.  I am thus disabling it until further notice.

[ Themes shipped with Emacs are not byte compiled (includes
  modus-operandi and modus-vivendi).  I have seen third-party themes do
  the same, such as Bozhidar Batsov's zenburn. ]

Thanks to Omar Antolín Camarena (@oantolin) for reporting the
aforementioned error.