Remove support for built-in vc-annotate
It has not been working properly for a long time now.  Colours are unset
and are not re-applied when switching between modus-operandi and
modus-vivendi.

Furthermore, the way vc-annotate-color-map intersects with
vc-annotate-background-mode puts us in an awkward spot: when the mode is
non-nil, the mapped values are used as backgrounds WITHOUT giving us the
chance to make the appropriate adjustments to the foreground (so we end
up with inaccessible colour combinations).  This means that we must fix
a problem which is not ours by overriding the user option of the
background altogether.  A theme outright disabling user options is bad
form.

Even documenting a user-level set of configurations will not suffice, as
the results are unreliable.  I tried the code which I copy further below
to test annotation with/without background, plus the change in values
when switching between modus-operandi and modus-vivendi.  Again, colours
are not updated properly (I know the buffer of M-x vc-annotate needs to
be generated again), as modus-operandi may retain the values set by
modus-vivendi or vice-versa.

Ultimately, I feel vc-annotate needs to be refactored to use ordinary
faces in ordinary ways.  Or, at least, not try to outsmart the
user/theme about the choice of colours.

Thanks to Philip Kaludercic for starting the thread about the
vc-annotate-background-mode which reminded me about this problem:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C875ylfxkgi.fsf@posteo.net%3E>.

The code I alluded to:

    (setq vc-annotate-background-mode nil)

    (defun my-modus-themes-vc-annotate ()
      ;; Actual values are for demo purposes
      (modus-themes-with-colors
        (if vc-annotate-background-mode
            (setq vc-annotate-background bg-alt
                  vc-annotate-color-map
                  `((20 .  ,red-intense-bg)
                    (40 .  ,red-subtle-bg)
                    (60 .  ,red-refine-bg)
                    (80 .  ,yellow-intense-bg)
                    (100 . ,yellow-subtle-bg)
                    (120 . ,yellow-refine-bg)
                    (140 . ,magenta-intense-bg)
                    (160 . ,magenta-subtle-bg)
                    (180 . ,magenta-refine-bg)
                    (200 . ,cyan-intense-bg)
                    (220 . ,cyan-subtle-bg)
                    (240 . ,cyan-refine-bg)
                    (260 . ,green-intense-bg)
                    (280 . ,green-subtle-bg)
                    (300 . ,green-refine-bg)
                    (320 . ,blue-intense-bg)
                    (340 . ,blue-subtle-bg)
                    (360 . ,blue-refine-bg)))
          (setq vc-annotate-background nil
                vc-annotate-color-map
                `((20 . ,red)
                  (40 . ,magenta)
                  (60 . ,magenta-alt)
                  (80 . ,red-alt)
                  (100 . ,yellow)
                  (120 . ,yellow-alt)
                  (140 . ,fg-special-warm)
                  (160 . ,fg-special-mild)
                  (180 . ,green)
                  (200 . ,green-alt)
                  (220 . ,cyan-alt-other)
                  (240 . ,cyan-alt)
                  (260 . ,cyan)
                  (280 . ,fg-special-cold)
                  (300 . ,blue)
                  (320 . ,blue-alt)
                  (340 . ,blue-alt-other)
                  (360 . ,magenta-alt-other))))))

    (add-hook 'modus-themes-after-load-theme-hook #'my-modus-themes-vc-annotate)