Revise all 'hi-*' faces using bespoke colours
Notice the use of :inverse-video as a means of preventing hl-line-mode
from overriding the background.  This is necessary due to the particular
requirements of this interface.

With default key bindings, type M-s h C-h to see relevant commands that
use these faces.

Thanks to Damien Cassou for the feedback on the mailing list:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C87bkuk1o24.fsf%40cassou.me%3E>.