REVISE current and lazy search colours
The previous combinations were not optimal, as the default green and
cyan are too close to each other.  Depending on factors such
environmental lighting and the quality of the monitor, green and cyan
may be indistinguishable.

The previous combinations was also not ideal when the user option
'modus-themes-deuteranopia' was set to a non-nil value: blue compared to
a subtle teal can also be problematic under certain circumstances.

The new colours (intense yellow for active matches and cool/cyan for
lazy ones) are complementary, meaning that there are naturally easy to
tell apart.  These specific hues are also well-suited for users with
red-green colour deficiency: yellow stays as-is, while the cool colour
becomes a bit more grey though remains distinct.  As such, we do not
need to run the helper function 'modus-themes--deuteran' to set the
style based on the value of 'modus-themes-deuteranopia'.

The new colours do not clash with the style of the 'match' face, nor
with the various permutations of the 'region' face (subject to the user
option 'modus-themes-region').