Remove 'bold' from active/lazy search
As commits c0c542a and ffe8cbe improve the state of affairs with regard
to the applicable colours, the inclusion of the 'bold' face is no longer
necessary.  Before we needed the 'bold' because the colours were not
sufficiently distinct.

Also, the bold weight makes paragraphs rendered in the 'variable-pitch'
face (or proportional fonts in general) jump around as the user moves
between the matches.  This was reported by Augusto Stoffel on the
mailing list:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C87sfnbswe9.fsf@gmail.com%3E>.