Set org-document-title via modus-themes-headings
The user option 'modus-themes-headings' now reads a level 0 heading.  It
accepts the same list of properties as all other levels.  Currently only
the Org document title is covered, but we may cover more faces if need
be.

Sample:

    (setq modus-themes-headings
          '((0 . (variable-pitch light (height 1.9)))
            (1 . (variable-pitch light (height 1.6) overline))
            (2 . (variable-pitch regular (height 1.4) overline))
            (3 . (variable-pitch regular (height 1.3) overline))
            (4 . (rainbow (height 1.2)))
            (5 . (rainbow (height 1.1)))
            (t . (variable-pitch extrabold (height 1.05)))))

Read the documentation of 'modus-themes-headings' for the technicalities.

Thanks to Rudolf Adamkovič for proposing the idea on the mailing list:
<https://lists.sr.ht/~protesilaos/modus-themes/%3Cm2y1x5tewl.fsf@me.com%3E>.

Given this change, I am also tweaking the default foreground value of
the org-document-title.  It is a bit more saturated, but remains close
to the spirit of the previous one.