Add more colour to date background in Org agenda
Makes it easier to discern where the highlight is, while it remains
close to the spirit of the previous design.  The idea is to not add too
much saturation here, because the buffer is already "busy" with lots of
highlights.

Thanks to Daniel Mendler for the feedback on the mailing list:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C3d8b1096-a7db-1e08-fefe-d39bed4a7ea3@daniel-mendler.de%3E>.