Remove support for centaur-tabs (unmaintained)
centaur-tabs has a bug where it cannot read the value of a face if it
uses the standard ':inherit' attribute.  I have sent a patch to fix it,
but have received no response since February:
<https://github.com/ema2159/centaur-tabs/pull/179>.

Relevant reports:

- <https://github.com/protesilaos/modus-themes/issues/30>
- <https://gitlab.com/protesilaos/modus-themes/-/issues/288>
- <https://github.com/protesilaos/modus-themes/issues/15>

I am happy to reinstate support for centaur-tabs as soon as the package
gets the maintenance it needs.