Remove support for paradox
I periodically test the packages we support to see if they have any
updates we need to cover but also to confirm that they work.  When I
tried paradox, it took over my C-c g binding which I have for Magit.  As
an Emacs user, I consider this an unacceptable transgression.

Looking at paradox's git repo, the project is not maintained.  If things
change, I am happy to reinstate support for it.