Make markdown-gfm-checkbox-face same as org-checkbox
This continues the work done in commit 2b0c31b.  The change comes from
the discussion on the mailing list where it became apparent that a bit
of colour is needed for such constructs (my suggestion for
'yellow-alt-other' was well received):
<https://lists.sr.ht/~protesilaos/modus-themes/%3Cm2fsi9cja4.fsf%40me.com%3E>

Thanks to Rudolf Adamkovič, Christian Tietze, and Karthik Chikmagalur
for their participation.