Make fill-column-indicator more noticeable
This is in response to private messages I received and this thread:
<https://lists.gnu.org/archive/html/help-gnu-emacs/2022-08/msg00255.html>.

The core idea is that the previous design is (1) "invisible" and (2)
prevents the customisation of display-fill-column-indicator-character.

I can address point 1, but point 2 puts us in an awkward spot as we
would then not have to use a background and a height value.  Not doing
so produces a dashed line by default, with the dashes further apart the
greater the line-spacing is.  It looks broken.