Use bespoke colors for the highlight-regexp faces
We cannot use our ordinary entries from each theme's palette as those
are not designed to handle such an edge case.  Those faces need to use
the inverse-video attribute which, in turn, affects the combinations
of color we can apply.  Our accented backgrounds are designed to
contrast well with our nominal main foreground values, whereas this
case demands colored backgrounds that contrast well with what would
normally be the main background color.  Furthermore, this case
requires that colored backgrounds are faithful to their name so that,
say, hi-yellow is yellowish.

Thanks to Kevin Kainan Li for the feedback on the mailing list, where
they informed me that the previous design was too dark/mute (and I
agreed with that assessment) and provided feedback on my samples:
<https://lists.sr.ht/~protesilaos/modus-themes/%3CCAMTq2Vp3Nnzv-i9wJdq4-OJ4X_QfWXySpUtAieBy0dgKLEOSBg%40mail.gmail.com%3E>.