Revert modus-themes-toggle obsolete
The 'toggle-theme' is not functionally equivalent to the command
'modus-themes-toggle' due to the optional arguments it accepts.  With
'toggle-theme' we are prompted to confirm loading the theme.  We also
get a second prompt that tells us how unsafe themes can be...  This is
consistent with how 'load-theme' works, but is not what the user of
'modus-themes-toggle' has come to expect.

Users who do not like to maintain a 'custom-file' are thus penalised
each time they invoke the command.  I will proceed with this
deprecation, if there is, say, a user option in Emacs that disables
those prompts each time a theme is loaded.  Basically, we need an
arrangement that just toggles themes without questions.

This commit reverts 59da1b3, 0260109.

Read the discussion on the mailing list with the participation of
Philip Kaludercic and Rudolf Adamkovič:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C877d116lh4.fsf%40posteo.net%3E#%3Cm2lepgrd8l.fsf@me.com%3E>.