Tweak telega code and pre faces
The 'telega-entity-type-code' is now subject to the value of the user
option 'modus-themes-markup'.  By default, it looks like Org
=verbatim= entities.  The 'telega-entity-type-pre' has the same style
as fenced code blocks in Markdown.

We cannot inherit from Org or another package, as there is no
guarantee it has been loaded by the time the relevant telega buffer is
displayed (in general, we do not let faces inherit from arbitrary face
groups).

Thanks to Pablo Stafforini for the feedback in issue 170 on the GitLab
mirror: <https://gitlab.com/protesilaos/modus-themes/-/issues/170#note_1130421067>.