Simplify telega-entity-type-pre face
The ':extend t' attribute does not work.  For it to have the desired
effect, the underlying fontification regexp must include the newline
character.  As a result, the face does not produce a code "block".
This new face is more appropriate in this regard.

Thanks to Pablo Stafforini for showing me screenshots of how it looks,
as I am not a telega/telegram user and cannot do this myself.  Done as
part of issue 170 on the GitLab mirror:
<https://gitlab.com/protesilaos/modus-themes/-/issues/170#note_1143975582>.