Make dev version 3.0.0 instead of 2.8.0
I plan to introduce some breaking changes (nothing major, but still)
so it is better to express this in the version number.  This will be
done in the interest of a better out-of-the-box experience for most
users.  I am timing it to coincide with the cut of the emacs-29
branch, which is expected in November.