BREAKING Make modus-themes-completions intense
Change the default value so that the background of the selected line
is a more intense gray than the previous design.  The idea is to
provide a more usable out-of-the-box experience for those who are not
aware that the themes are customisable.

To restore the old default with the more subtle gray, add this to the
init file before loading the theme:

    (setq modus-themes-completions nil)

As with all user options for themes, changes are applied upon theme
reload.