BREAKING Make modus-themes-hl-line intense
Change the default value so that the background of the currently
highlighted line with 'hl-line-mode' is a more intense gray than the
previous design.  The idea is to provide a more usable out-of-the-box
experience for those who are not aware that the themes are
customisable.

To restore the old default with the more subtle gray, add this to the
init file before loading the theme:

    (setq modus-themes-hl-line nil)

As with all user options for themes, changes are applied upon theme
reload.