BREAKING Make modus-themes-fringes visible
Change the default value so that the background is a subtle gray.  The
idea is to provide a more usable out-of-the-box experience for those
who are not aware that the themes are customisable.

To restore the old default with the invisible fringes, add this to the
init file before loading the theme:

    (setq modus-themes-fringes nil)

As with all user options for themes, changes are applied upon theme
reload.