Make olivetti-fringe use the main background
The previous style did not actually hide the fringes, as olivetti
would still fall back to reading the generic fringe face.

The intent of this change is to make the olivetti-fringe invisible.

Thanks to Matthias Fuchs for reporting the issue that helped me track
this problem.  This was done in issue 46 on the GitHub mirror:
<https://github.com/protesilaos/modus-themes/issues/46>.