Add support for agda2-mode
This is based on merge request 44 by Xinglu Chen on the GitLab mirror
from 2021-06-16:
<https://gitlab.com/protesilaos/modus-themes/-/merge_requests/44>.
That contribution could not be merged, as it exceeded the ~15 line
limit and thus required copyright assignment to the Free Software
Foundation.  The author became inactive soon thereafter despite their
stated intention to do the paperwork.

Thanks to Laith Bahodi for informing me about this in issue 48 on the
GitHub mirror: <https://github.com/protesilaos/modus-themes/issues/48>.

In principle, the agda project should be inheriting from font-lock
faces instead of hardcoding their own colour values: those do not play
nicely with themes.  If anyone familiar with agda reads this, I
encourage you to work with upstream to make things better for all
themes.  Also consider that adga cannot be tested without an
established setup for Haskell, which adds to the burden of maintaining
those faces at the theme level (I could not test them and only went by
screenshots and prior experience).