Deprecate user options for colour overrides
This model was too difficult for users:

* One had to overwrite too many colours to get the desired results.
* It was not possible to change colour mappings so, for example,
  keywords would always be purple.

The purpose of the rewrite for version 3.0.0 is to abstract the colour
palette and create colour mappings of "semantic colours".  The user
can then edit an existing theme or derive one.  We will reduce the
number of entries in the palette and rename colours in the interest of
providing self-explanatory code.