Make mode-line-highlight always have a border
This is because the background colour it is using now may not be
discernible against an accented mode line.

The use of ':box' ensures that the padding of the highlight is always
the same as that of the mode line, whereas before we had nested
padding values (a padded box inside a padded box).