Reinstate functionality of auto theme reload upon option switch
The 'modus-themes-custom-auto-reload' is the successor to the
now-deprecated 'modus-themes-inhibit-reload'.  The new name is more
descriptive and easier to reason about.

The functionality is now enabled by default.  I am using it with
'setopt' (Emacs 29) and think it is a user-friendly feature.

Thanks to Philip Kaludercic for the feedback on the mailing list:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C87h6y3gyxx.fsf%40posteo.net%3E>.

Philip is the author of the original 'modus-themes-inhibit-reload' and
the relevant setup.  I just made some minor tweaks for this "version
4" rewrite of the themes.  Any mistakes are my own.