Add [tentative] support for menu, scroll-bar, tool-bar faces
Emacs does not have control over those GUI elements as they are drawn
by the underlying toolkit.  I built Emacs with Lucid and the 'menu'
and 'tool-bar' responded to those faces.  For the 'scroll-bar', Lucid
seems to only affect its background.

On GTK builds, the 'menu' and 'tool-bar' faces have no effect.  The
'scroll-bar' seems to affect both the background and foreground of the
given element.

Consider this change subject to review.