Make 'compilation-warning' also inherit 'modus-themes-bold'
This was always the intention for all three of
'compilation-{error,info,warning}'.  They will use a bold weight when
the user option 'modus-themes-bold-constructs' is non-nil.

Thanks to Manuel Giraud for informing me about the inconsistency in
'compilation-warning'.  It was done via a private channel and the
information is shared with permission.