Add user option modus-themes-disable-other-themes
In the past, we used to disable all other themes while loading a Modus
theme.  I revised this for version 4 because I learnt that some users
use "themes" as bundles of arbitrary configurations.  This was
supported by the fact that the Modus themes are built into Emacs so we
need to accommodate every workflow that Emacs supports.

With this addition, I am reverting to the old way of doing things:
loading a Modus theme should disable all others.  Users who need to
run multiple themes can set this option to a nil value.

Thanks to Tony Zorman for the feedback on the mailing list:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C874jtew0gp.fsf%40hyperspace%3E#%3C87bkne84d2.fsf@hyperspace%3E>.