Make Org block delimiter text more intense when "gray background"
This changes the colour of the #+begin_src line (and corresponding end
line) to be more intense when the user option 'modus-themes-org-blocks'
is set to the value 'gray-background'.

That style makes the inside of the block have a gray background and
the begin/end lines to have a more intense gray.  So those lines
cannot retain gray text that is there without the background.