Add function modus-themes-get-color-value and document it
I already had the functionality internally, but it is good to expose
it to users.  I am inspired to do so after checking Sacha Chua's blog
post "Making highlight-sexp follow modus-themes-toggle".[1] Sacha is
using a private function from the themes, which will be problematic if
I ever move things around.

[1] <https://sachachua.com/blog/2023/01/making-highlight-sexp-follow-modus-themes-toggle/>