Make diff-mode also use foreground values
The 'diff-font-lock-prettify' user option is set to retain the
underlying highlights.  While this is "prettified", it is not good for
legibility.

The added foregrounds apply to the text of the diffs that would
otherwise use the 'default' face.  This way we make the added/removed
easier to spot.  Also, adding those foregrounds allows us to use
palette overrides to implement a "foregrounds only" style for diffs (I
will update the manual accordingly).