Acknowledge Gautier Ponsinet for commit 55d1651
The commit in question introduces the org-agenda-calendar-daterange
face, which is part of the yet-to-be-released Org version 9.7.  The
change is below the ~15 line limit and thus does not require copyright
assignment to the Free Software Foundation.

I received the patch via a private channel.  Thanks to Gautier for
doing this!

Also see commit 9869041 that introduces the semantic colour mapping in
each theme's palette.