Make Jinx use a warning instead of an error
This is because of how the package works: it automatically highlights
misspellings in the visible portion of the buffer.  There are cases
where this results in a very intense presentation, which can be
distracting.  We want to reduce the overall intensity and not draw too
much attention to those highlights.