Revise bg-hover hue in all themes except tritanopia
The previous colour was not sufficiently distinct from bg-completion.
This would make it hard to spot the difference while, for example,
using 'vertico-mode' in tandem with 'vertico-mouse-mode'.

Same principle for the difference between the mouse hover and lazy
isearch highlights (e.g. in Dired or Occur buffers).

Changing the hue here follows the same principle that underpinned the
redesign of the grey backgrounds for version 4 of the project:
depending on hardware capabilities, colour reproduction may not be
optimal, so we need to be more considerate with the choice of colour
values.

The tritanopia themes are not affected by this, as they already use
highly distinct hues.

Thanks to Daniel Mendler[1] for bringing this matter to my attention
and for testing the proposed alternatives.  This was done via a
private channel and the information is shared with permission.

[1] <https://github.com/minad>