Remove "hover" overrides from "faint" preset
Those contradict the changes done in commit 4e331d2.  In short, we
want a different colour for the hover effects so that it is not
mistaken for other elements.

Thanks to Daniel Mendler[1] for bringing this matter to my attention.
This was done via a private channel and the information is shared with
permission.

[1] <https://github.com/minad>