Add semantic colour mappings for terminal emulators
These are used by ansi-term, vterm, and the like.  The idea is to
empower users to differentiate background and foreground values.  By
convention, terminal emulators use the same value for both background
and foreground, although this is not optimal with high contrast themes
because what works as a foreground does not necessarily look nice as a
background.

The default values of the new mappings retain the prior state, just to
not break existing configurations.

Thanks to Tony Zorman for reporting the bug that provided the impetus
for this change:
<https://lists.sr.ht/~protesilaos/modus-themes/%3C87fs4wforf.fsf%40hyperspace%3E>.