Style the 'consult-preview-insertion' face
The default style inherits from the 'region' face. This is fine in
some cases, though not when the user calls 'consult-outline' in a
buffer with hidden text (e.g. Org folded headings). There the
background is applied to the entire heading's surface area, which is
too distracting while cycling through completion candidates and
inspecting the preview.