Simplify check for Org block backgrounds; make them have bg with "tinted" option
Now when the user sets the option 'modus-themes-org-blocks' to a
non-nil value, the themes will always apply a background. It will be
gray by default unless the value is 'tinted-background'.

Before, there was no fallback background for quote and verse blocks
when the user was setting 'modus-themes-org-blocks' to the
'tinted-background'. We always want to have a background in that case.
The quote/verse blocks use gray at all times because we cannot modify
them via the 'org-src-block-faces'.

Thanks to Lennart C. Karssen for bringing this matter to my attention
in issue 313 on the GitLab mirror:
<https://gitlab.com/protesilaos/modus-themes/-/issues/313>