Remove some needless :set functions
Our :set function is designed to reload the theme, with the intent to
apply whatever change to the visuals. But this is not relevant for
user options that have no direct effect on the appearance of the
themes.

This is done as part of a discussion with shipmints in issue 118:
<https://github.com/protesilaos/modus-themes/issues/118>.