Simplify modus-themes--retrieve-palette-value
If the value is not contained in the palette, the recursive call to
modus-themes--retrieve-palette-value will return unspecified anyway,
thanks to the alist-get at the start of the function.

This saves a list allocation per recursive call, at the expense of
an additional funcall.

An alternative that avoids both the list allocation and the extra
funcall is to use (assq value palette) instead, at the expense of
more (and repetitive) typing.